Refactor get_single_alleles method
get_single_alleles method was reading a csv and parsing the index to get genes with single alleles. moved the file reading to __init__ so that it can be called from outside the class by just providing the df_alleles dataframe.
It also made unit testing for this function easier because we just have to pass a dataframe. Had to update unit testing function find_junction_init_org to patch the new file read function in __init__.