Change get_single_alleles params to any iterable
The get_single alleles previously took a dataframe as input and went through the gene names in its index to find genes with single alleles. Now it can take in any iterable with gene names which makes its function very clear.
Additionally, this allows us to now use the fasta file itself as the source of the gene name so we no longer need the allele by genome df to calculate the junctions. The test_junction was modified to account for this change in
parameter.