Add calc_kmer function to FindJunction class
This new function lets you use ntCard algorithm to get kmer frequency distributions in your fasta file. Overall the function is organized similarly to calculate junction in that it uses _yield_group_seq to get sequences for gene cluster and maps
that cluster to new process that calculates the kmer frequency distribution for the cluster. While its working, its takes ~2.5 seconds per cluster is not taking advantage of multi-processing. Not sure why that is. Additionally,
I had to change the _yield_gene_cluster function by adding a 'group' (default=False) param. Previously, because of requirement for twopaco, this function  wrote a new fasta file for each of the allele sequences. Now, if you set group=True, it will write all the allele
sequences for a gene cluster in a single file. The grouped sequence file works better with ntCards.