Add calc_gldist function
The function calcuates the range of gene lengths in different gene clusters i.e. give max(gene_length) - min(gene_length) for each gene cluster in the fasta file. While testing this, I found a bug in _yield_gene_cluster
function where it was overwriting the fasta files when group=True. Switched the write mode from 'w' to 'a+' to fix this.