Implement new way to build sparse matrix from junction file
The old of way making sparse matrix by reading one line at a time of the junction file took >10 hours even in small staph database. Rewrote new functions to do this in ~5mins. The function produces an output, still need to implement tests to make sure that the outputs are numerically correct. Also got rid of the old slow function to make the sparse matrix so that test currently fails on this commit. Will fix on next commit.